Project Final GarudaHQ
Mansion is a tech company that engaged in education and digital marketing advertising. Providing services, such as business class to improve the soft skills of people in micro, small and medium enterprises (MSME) sector. And also digital advertising services that are intended to help large and small MSMEs to be able to market products, form product images, and develop their business. Mansion is engaged in the form of a website, and presents interesting features and of course helps users according to their needs